Create README.md
Tour the world from your phone. Press and hold for a second on the map to add a pin for a location. Then tap the pin to get a collection view of photos, downloaded from flickr, for that location. The Virtual Tourist project was part of the Udacity Nanodegree program.